requirements.txt
-e git://github.com/tymofij/django.git@stable/1.3.x-py27#egg=django

celery==3.0.11
chardet==2.1.1
django-ajax-selects==1.1.4
django-appconf==0.5
django-authority==0.4
django-celery==3.0.11
django-filter==0.5.3
django-guardian==1.0.4
django-kombu==0.9.4
django-picklefield==0.2.1
django-social-auth==0.7.8
django-tagging-autocomplete==0.3.1
django-userena==1.1.2
easy-thumbnails==1.1
hiredis==0.1.1
httplib2==0.7.6
markdown==2.2.1
mock
Pillow
polib==0.6.3
pygooglechart
pysolr==2.1.0
redis==2.7.1
requests==0.14.2
south==0.7.3
BeautifulSoup==3.2.0
django-staticfiles==1.2.1

./vendor/contact_form-0.3.tar.gz
./vendor/django-piston-0.2.3.tar.gz
./vendor/django-sorting-0.1.tar.gz
./vendor/django-threadedcomments-0.9.tar.gz
./vendor/userprofile-0.7-r422-correct-validation.tar.gz
./vendor/django-addons-0.6.6.tar.gz

-e git://github.com/tymofij/kombu.git@2.5-py27#egg=kombu

-e git://github.com/jezdez/django_compressor.git@ee16af610d677a0a9b63602da6558f601f66287c#egg=django_compressor-dev
-e git://github.com/jkal/django-notification.git@83b83551df5e4b5c83b20f4db9cc0efdcff40a30#egg=django_notification-dev
-e git://github.com/jkal/django-pagination.git@15515fcc1fa1afc150a3a6565efadcd3700f83e7#egg=django_pagination-dev
-e git://github.com/mpessas/django-bulk.git@e3d606186d64d854857060d88e57bca9948d61f1#egg=django_bulk-dev
-e git://github.com/mpessas/django-tagging.git@a19a54bad767c29595a3e0c39e6331cf91f36bd7#egg=django_tagging-dev
-e git://github.com/mpessas/python-Levenshtein.git@71c1ce1b430435c6265c0f558d4b7aba49100d19#egg=python_Levenshtein-dev
-e git://github.com/toastdriven/django-haystack.git@3289ab8bb410321a94608a13e2e527ddbd8f1a7e#egg=django_haystack-dev

